% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinecorrplot.R
\name{baselinecorrplot}
\alias{baselinecorrplot}
\title{Plot the result of optimal baseline correction}
\usage{
baselinecorrplot(mscorr, msorig = NULL, i = 1L, addorig = FALSE,
  methodused = "SNIP", ...)
}
\arguments{
\item{mscorr}{a list of MALDIquant::massSpectrum objects after baseline
correction}

\item{msorig}{an optional list of MALDIquant::massSpectrum objects before
baseline correction (defaults to NULL)}

\item{i}{an integer (defaults to 1) selecting the cell number. Cannot be
larger than the number of cells in the mscorr list}

\item{addorig}{a boolean indicating wether or not to add the spectrum before
baseline correction to the plot. If the list containing this data is not
supplied, an error will be thrown (in the future this should be a warning).
Defaults to FALSE}

\item{methodused}{a string indicating which method was used for baseline
correction. Only used in the plot title. Defaults to "SNIP".}

\item{...}{additional parameters passed on to plot}
}
\description{
This function plots a selected baseline-corrected Raman spectrum.
Optionally, the original spectrum can be added (for reference). At the moment
the function assumes that the cells are in the same order in the baseline-
corrected and original data.
}
\examples{
## Short example

# Load hyperSpec object
data("mdqs")
data("mass.spectra.baseline.corr")

# Extract wavelengths and plot
mdqs.trim <- trim(mdqs, range=c(600, 1800))
baselinecorrplot(mass.spectra.baseline.corr,mdqs.trim,i=2,addorig=TRUE)

}
