% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluscutplot.R
\name{cluscutplot}
\alias{cluscutplot}
\title{Plot clusters on hclust object at given height.}
\usage{
cluscutplot(dend, h = 0.75, bordercol = "red", ...)
}
\arguments{
\item{dend}{a hclust object}

\item{h}{cutoff distance for the clusters to be plotted (defaults to 0.75)}

\item{bordercol}{the color of the border}

\item{...}{other parameters passed on to plot.hclust}
}
\value{
a matrix with the assigned clusters for each leaf (in the case of Raman data each cell)
}
\description{
An easy exploratory way to draw clusters at a given height on a dendrogram.
}
\examples{
## Short example
data("SCA.similarity")
dendrogram <- hclust(SCA.similarity, method="ward.D2")
cluscutplot(dendrogram)
}
