% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange_shared_legend.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Function for creating a common legend for 2 ggplot2 figures.}
\usage{
grid_arrange_shared_legend(..., ncol = length(list(...)), nrow = 1,
  position = c("bottom", "right"))
}
\arguments{
\item{...}{ggplot2 objects which could have a common legend}

\item{ncol}{number of columns in the grid (defaults to the number of plots)}

\item{nrow}{number of rows in the grid (defaults to 1)}

\item{position}{character vector (of length 2) defining the position of the legend. Defaults to bottom and right.}
}
\description{
Function for creating a common legend for 2 ggplot2 figures.
}
