% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs2mq.R
\name{hs2mq}
\alias{hs2mq}
\title{Convert a hyperSpec object toa  MassSpectrum object}
\usage{
hs2mq(x)
}
\arguments{
\item{x}{hyperSpec object
## Short example

# Load hyperSpec object
data("spx.all")

# Convert to MassSpectrum object
mdqs <- lapply(spx.all,hs2mq)}
}
\description{
This function converts a hyperSpec::hyperSpec objects (default of Raman spectra) to a
MALDIquant::MassSpectrum object for downstream processing.
}
