% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_coll_curve.R
\name{hs_coll_curve}
\alias{hs_coll_curve}
\title{Collection curves for phenotypic heterogeneity measurements}
\usage{
hs_coll_curve(
  hs.x,
  intervals = 10,
  nboot = 10,
  peak_detection = FALSE,
  peak_window = 20,
  peak_method = c("MAD"),
  plot_fig = TRUE
)
}
\arguments{
\item{hs.x}{HyperSpec object}

\item{intervals}{Decide in which sample size intervals you want to inspect the curves.
Defaults to 10 cells.}

\item{nboot}{Number of bootstraps to condut for each sample size. Defaults to 10.}

\item{peak_detection}{Should peak detection be used instead of raw spectra for Hill calculations? Defaults to FALSE}

\item{peak_window}{Peak windows size for a peak to be considered a signal. Defaults to 20.}

\item{peak_method}{Which peak detection method should be used (requires peak_detection == TRUE).
Options are "MAD" and "SuperSmoother"}

\item{plot_fig}{Should figures of collectors curves be shown? Defaults to TRUE.}
}
\description{
Calculates collection curves that describe the sensitivity of the diversity metrics to sample size.
This can be used to check whether the Raman spectra in their totality cover the phenotypic heterogeneity of the sample.
}
\examples{
# Short example
data("hs_example")

# Preprocess
hs_example <- hs_preprocess(hs_example)

# Calculate metrics
hs_coll_curve(hs_example, intervals = 3, nboot = 10, plot_fig = FALSE)
}
