% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_conv_mq.R
\name{hs_conv_mq}
\alias{hs_conv_mq}
\title{Convert a hyperSpec object toa  MassSpectrum object}
\usage{
hs_conv_mq(hs.x)
}
\arguments{
\item{hs.x}{hyperSpec object}
}
\description{
This function converts a hyperSpec::hyperSpec objects (default of Raman spectra) to a
MALDIquant::MassSpectrum object for downstream processing.
}
\examples{
## Short example

# Load hyperSpec object
data("hs.example")

# Convert to MassSpectrum object
mdqs <- hs_conv_mq(hs.example)
}
