% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_example.R
\docType{data}
\name{hs_example}
\alias{hs_example}
\title{Raman spectrometry test single-cell data of GFP expressing yeast culture}
\format{
Hyperspec object
}
\usage{
hs_example
}
\description{
Raman spectrometry test single-cell data of GFP expressing yeast culture
}
\keyword{datasets}
