% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_hclust.R
\name{hs_hclust}
\alias{hs_hclust}
\title{Perform hierarchical clustering on Raman spectra}
\usage{
hs_hclust(hs.x, dist_method = "bray", ...)
}
\arguments{
\item{hs.x}{hyperSpec object}

\item{dist_method}{Dissimilarity index, partial match to "manhattan", "euclidean", "canberra", "clark",
"bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn",
"mountford", "raup", "binomial", "chao", "cao" or "mahalanobis".}

\item{...}{Parameters to pass on to hclust().}
}
\description{
Perform hierarchical clustering on Raman spectra
}
\examples{
## Short example

# Load hyperSpec object
data("hs.example")

# Convert to MassSpectrum object
hs.x.proc <- hs_preprocess(hs.x)

hclust_obj <- hs_hclust(hs.x.proc)
}
