% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_hclust_cutoff.R
\name{hs_hclust_cutoff}
\alias{hs_hclust_cutoff}
\title{Plot clusters on hclust object at given height.}
\usage{
hs_hclust_cutoff(dend, h = 0.75, bordercol = "red", ...)
}
\arguments{
\item{dend}{a hclust object}

\item{h}{cutoff distance for the clusters to be plotted (defaults to 0.75)}

\item{bordercol}{the color of the border}

\item{...}{other parameters passed on to plot.hclust}
}
\value{
a matrix with the assigned clusters for each leaf (in the case of Raman data each cell)
}
\description{
An easy exploratory way to draw clusters at a given height on a dendrogram.
}
\examples{
# Load hyperSpec object
data("hs.example")

# Convert to MassSpectrum object
hs.x.proc <- hs_preprocess(hs.x)

# cluster cells
hclust_obj <- hs_hclust(hs.x.proc)

# Draw cutoff
hs_hclust_cutoff(hclust_obj)
}
