% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_phenoRam.R
\name{hs_phenoRam}
\alias{hs_phenoRam}
\title{Calculate the phenotypic heterogeneity indices for each cell}
\usage{
hs_phenoRam(
  hs.x,
  preprocess = FALSE,
  smooth = FALSE,
  align = FALSE,
  path = NULL,
  trim.range = c(400, 1800),
  pattern = ".spc",
  niter = 10,
  return_spectra = TRUE,
  peak_detection = FALSE,
  peak_window = 20,
  peak_method = c("MAD")
)
}
\arguments{
\item{hs.x}{HyperSpec object}

\item{preprocess}{Preprocess? TRUE/FALSE. Defaults to FALSE.}

\item{smooth}{Option for preprocessing. Defaults to FALSE.}

\item{align}{Align option for preprocessing. Defaults to FALSE.}

\item{path}{path where Raman data is located, only useful if preprocess==TRUE}

\item{trim.range}{Wavenumber range to trim from raw spectra. Only used if preprocess==TRUE}

\item{pattern}{(POSIX) regular expression used to select files (defaults to ".spc"), only used if preprocess==TRUE
Options are "MAD" and "SuperSmoother"}

\item{niter}{Number of iteration Option for preprocessing. Defaults to FALSE.}

\item{return_spectra}{Should hyperSpec object be returned as final object? Defaults to TRUE.}

\item{peak_detection}{Should peak detection be used instead of raw spectra for Hill calculations? Defaults to FALSE}

\item{peak_window}{Peak windows size for a peak to be considered a signal. Defaults to 20.}

\item{peak_method}{Which peak detection method should be used (requires peak_detection == TRUE).}
}
\description{
Calculate the phenotypic heterogeneity indices for each cell
}
\examples{
# Short example
data("hs_example")

# Preprocess
hs_example <- hs_preprocess(hs_example)

# Calculate metrics
hs_phenoRam(hs.x = hs_example, preprocess = TRUE, peak_detection = TRUE)
}
