% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_preprocess.R
\name{hs_preprocess}
\alias{hs_preprocess}
\title{Calculate the phenotypic heterogeneity indices for each cell}
\usage{
hs_preprocess(
  hs.x,
  smooth = FALSE,
  align = FALSE,
  path = NULL,
  pattern = ".spc",
  trim.range = c(400, 1800),
  niter = 10
)
}
\arguments{
\item{hs.x}{Available hyperspec object}

\item{smooth}{Smoothing TRUE/FALSE}

\item{align}{Algn spectra TRUE/FALSE}

\item{niter}{Number of iterations.}

\item{peak_detection}{d}

\item{peak_window}{d}

\item{peak_method}{d}

\item{d}{d}
}
\description{
Calculate the phenotypic heterogeneity indices for each cell
}
\examples{

## Short example

# Load hyperSpec object

}
