% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_resample.R
\name{hs_resample}
\alias{hs_resample}
\title{Resampling function for hyperspec objects}
\usage{
hs_resample(hs.x, sample = 0, replace = FALSE, seed = NULL)
}
\arguments{
\item{hs.x}{hyperSpec object.}

\item{sample}{Desired sample size. Defaults to minimum sample size.}

\item{replace}{Do you want to resample with or without replacement? Defaults to FALSE, which is without replacement.}

\item{seed}{Set seed for reproducibile sampling if that is necessary. Defaults to NULL.}
}
\value{
Hyperspec object
}
\description{
This function resamples the hyperspec object to an equal number of cells/spectra. Options are to resample
with or withour replacement.
}
\examples{
# Short example
data("hs_example")

# Preprocess
hs_example <- hs_preprocess(hs_example)

# Resample
hs_example_r <- hs_resample(hs_example, sample = 10, replace = TRUE)
}
\keyword{hyperSpec}
\keyword{resampling,}
