% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_RF_pred.R
\name{hs_RF_pred}
\alias{hs_RF_pred}
\title{Train random forest model on hyperspec object}
\usage{
hs_RF_pred(hs.x, model, ...)
}
\arguments{
\item{hs.x}{Hyperspec object on which to perform predictions}

\item{model}{Model to use for predictions. Output from hs_RF.}

\item{...}{additional parameters passed on to stats::predict}
}
\description{
Train random forest model on hyperspec object
}
\examples{
# Short example
data("hs_example")

# Preprocess
hs_example <- hs_preprocess(hs_example)

# Mock-up metadata
mock_meta <- data.frame(Spectrum_ID = rownames(hs_example@data$spc),
group = factor(c(rep(1,30),rep(2,34))))

# Calculate metrics
hs.RF <- hs_RF(hs.x = hs_example, metadata = mock_meta, spectrumID_col= "Spectrum_ID",
 target_var = "group")

 # Trained model
 print(hs.RF[[1]])

 # Confusion matrix
 print(hs.RF[[2]])

 # Variable importance metric
 caret::varImp(hs.RF[[1]])

 # Perform predictions
 hs_RF_pred(hs.x = hs_example, model = hs.RF[[1]])
}
