% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_SCA_conv_itol.R
\name{hs_SCA_conv_itol}
\alias{hs_SCA_conv_itol}
\title{Calculate the spectral conrast angle (SCA) and generate an iTol-compatible tree}
\usage{
hs_SCA_conv_itol(
  x,
  Factor1 = c("LB", "NB", "LB"),
  Factor1Name = "Medium",
  Factor2 = c("rep1", "rep2", "rep3"),
  Factor2Name = "Replicate",
  clusMethod = "ward.D2"
)
}
\arguments{
\item{x}{a hyperSpec object containing the cells in the rows, and wavenumbers (features) in columns}

\item{Factor1}{a character vector containing treatment levels (defaults to different media)}

\item{Factor1Name}{a string naming Factor 1 (defaults to "Medium")}

\item{Factor2}{a character vector containing "treatment" levels (defaults to different replicates)}

\item{Factor2Name}{a string naming Factor 2 (defautls to "Replicate")}

\item{clusMethod}{a string showing which stats::hclust linkage method should be chosen}
}
\value{
an ape::phylo object based upon hclust of SCA distances of the hyperSpec object's cells
}
\description{
The SCA measures the angle between two vectors corresponding to closely
 related spectra to measure whether they are the same or not
 (Wan *et al.*, 2002). This function returns an iTol (Letunic & Bork, 2016)
}
\references{
Wan, K. X., Vidavsky, I., & Gross, M. L. (2002). Comparing similar spectra: from similarity index to spectral contrast angle. Journal of the American Society for Mass Spectrometry, 13(1), 85-88.

Letunic, I., & Bork, P. (2016). Interactive tree of life (iTOL) v3: an online tool for the display and annotation of phylogenetic and other trees. Nucleic acids research, 44(W1), W242-W245.
}
