% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_SCAdiss.R
\name{hs_SCAdiss}
\alias{hs_SCAdiss}
\title{Calculate the spectral contrast angle (SCA) between all cells in a hyperSpec object}
\usage{
hs_SCAdiss(x)
}
\arguments{
\item{x}{a hyperSpec object containing the cells in the rows, and wavenumbers (features) in columns}
}
\value{
a distance object containing the SCA between each cell
}
\description{
The SCA measures the angle between two vectors corresponding to closely
 related spectra to measure whether they are the same or not
 (Wan *et al.* 2002). This function returns a dist object based on the SCA
 for downstream processing.
}
\examples{
# Load hyperSpec object
data("hs_example")

# Convert to MassSpectrum object
hs.x.proc <- hs_preprocess(hs_example)

# cluster cells
disst <- hs_SCAdiss(hs.x.proc)
}
\references{
Wan, K. X., Vidavsky, I., & Gross, M. L. (2002). Comparing similar spectra: from similarity index to spectral contrast angle. Journal of the American Society for Mass Spectrometry, 13(1), 85-88.
}
