% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_tidy_filenames.R
\name{hs_tidy_filenames}
\alias{hs_tidy_filenames}
\title{hs_tidy_filenames function}
\usage{
hs_tidy_filenames(
  hs.x,
  remove_pattern = "path",
  rename_df = NULL,
  old_name_col,
  new_name_col
)
}
\arguments{
\item{hs.x}{HyperSpec object}

\item{remove_pattern}{What should be adjusted in filenames.
"path" means remove the absolute path of the file and only keep the filename
"rename" means that you supply a dataframe with matching old and new filenames.
In this case you must supply a dataframe in rename_df and designate the old filename
column (old_name_col) and new filename column (new_name_col).}

\item{rename_df}{dataframe containing old and new filenames for the hyperspec objects}

\item{old_name_col}{column in rename_df with old filenames}

\item{new_name_col}{column in rename df with new filenames}
}
\description{
Function to tidy/clean filenames in hyperspec objects (i.e. remove absolute path in name, and/or replace filenames).
}
\examples{
data("hs_example")

# Remove path from spectral ID
hs_tidy_filenames(hs_example, remove_pattern = "path")

}
\keyword{hyperspec}
\keyword{rename}
