% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_tsne.R
\name{hs_tsne}
\alias{hs_tsne}
\title{Run tSNE on a hyperspec object}
\usage{
hs_tsne(hs.x, ...)
}
\arguments{
\item{hs.x}{hyperSpec object or alternatively distance matrix}

\item{...}{Parameters to pass on to tsne().}
}
\description{
Run tSNE on a hyperspec object
}
\examples{
## Short example

# Load hyperSpec object
data("hs.example")

# Convert to MassSpectrum object
hs.x.proc <- hs_preprocess(hs.x)

df.tsne <- hs_tsne(hs.x.proc)
}
