% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_type.R
\name{hs_type}
\alias{hs_type}
\title{Perform clustering of spectra using partitioning around medoids (PAM)}
\usage{
hs_type(hs.x, PCA.var = NULL, ...)
}
\arguments{
\item{hs.x}{hyperSpec object}

\item{PCA.var}{Variance threshold for PCA preprocessing prior to clustering.
Defaults to NULL which means that no PCA will be performed.}

\item{...}{Parameters to pass on to pam().}
}
\description{
Number of clusters is optimized based on Silhouette index.
}
\examples{
## Short example

# Load hyperSpec object
data("hs_example")

# Convert to MassSpectrum object
hs.x.proc <- hs_preprocess(hs_example)

hclust_obj <- hs_type(hs.x.proc, PCA.var = 0.8)
}
