% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalplot.r
\name{intervalplot}
\alias{intervalplot}
\title{Plot a range of wavelengths and their intervals}
\usage{
intervalplot(trimmedwl)
}
\arguments{
\item{trimmedwl}{a numeric vector of wavelengths (trimmed to the biological relevant part)}
}
\description{
This function plots the intervals between wavelengths to assess whether
or not they were constant.
}
