% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_stats.R
\name{model_fit_stats}
\alias{model_fit_stats}
\title{Extract model fit statistics from a linear model}
\usage{
model_fit_stats(linear.model)
}
\arguments{
\item{linear.model}{an lm-type object}
}
\value{
A data.frame containing the R squared, adj. R squared, pred. R squared,
    and predictive residuals
}
\description{
This function extracts the R squared, adjusted R squared, predicted R squared,
and predictive residuals from a linear model.
}
\examples{
## Short example

}
