% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq2hs.R
\name{mq2hs}
\alias{mq2hs}
\title{Convert a MassSpectrum object to a hyperSpec object}
\usage{
mq2hs(x)
}
\arguments{
\item{x}{hyperSpec object}
}
\description{
This function converts a list of MALDIquant::MassSpectrum  objects to a
hyperSpec::hyperSpec object for downstream processing. It assumes all
MassSpectrum objects have the same wavelengths as the first entry in the list.
}
\examples{
## Short example

# Load and pretransform MassSpectrum object
data("mass.spectra.baseline.corr")
mass.spectra.baseline.corr <- wlcutter(mass.spectra.baseline.corr)
mq.norm <- calibrateIntensity(mass.spectra.baseline.corr, method="TIC",range=c(600, 1800))
# Convert to hyperSpec object
hs.norm <- mq2hs(mq.norm)
}
