% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_conv_hs.R
\name{mq_conv_hs}
\alias{mq_conv_hs}
\title{Convert a MassSpectrum object to a hyperSpec object}
\usage{
mq_conv_hs(mq.x)
}
\arguments{
\item{mq.x}{list of MALDIquant objects}
}
\description{
This function converts a list of MALDIquant::MassSpectrum  objects to a
hyperSpec::hyperSpec object for downstream processing. It assumes all
MassSpectrum objects have the same wavelengths as the first entry in the list.
}
\examples{
# Load hyperSpec object
data("hs.example")

# Convert to MassSpectrum object
mq.x <- hs_conv_mq(hs.example)

# Convert to hyperspec object
hs.x <- mq_conv_hs(mq.x)
}
