% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoRam.R
\name{PhenoRam}
\alias{PhenoRam}
\title{Calculate the phenotypic heterogeneity indices for each cell}
\usage{
PhenoRam(
  hs.x,
  preprocess = FALSE,
  smooth = FALSE,
  align = FALSE,
  path = NULL,
  trim.range = c(400, 1800),
  pattern = ".spc",
  niter = 10,
  return_spectra = TRUE,
  peak_detection = TRUE,
  peak_window = 20,
  peak_method = c("MAD")
)
}
\arguments{
\item{hs.x}{d}

\item{preprocess}{d}

\item{smooth}{d}

\item{align}{d}

\item{niter}{d}

\item{return_spectra}{d}

\item{peak_detection}{d}

\item{peak_window}{d}

\item{peak_method}{d}

\item{d}{d}
}
\description{
Calculate the phenotypic heterogeneity indices for each cell
}
\examples{

## Short example

# Load hyperSpec object

}
