% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_r_squared.R
\name{pred_r_squared}
\alias{pred_r_squared}
\title{Calculated predicted R squared from a linear model}
\usage{
pred_r_squared(linear.model)
}
\arguments{
\item{linear.model}{an lm-type object}
}
\value{
A numeric with the predicted R squared
}
\description{
This function calculates the predicted R squared from a linear model
}
