% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRESS.R
\name{PRESS}
\alias{PRESS}
\title{Calculated predicted residuals from a linear model}
\usage{
PRESS(linear.model)
}
\arguments{
\item{linear.model}{an lm-type object}
}
\value{
A numeric with the predicted residuals
}
\description{
This function calculates the predicted resiiduals from a linear model
}
