% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mq_plot.R
\name{ramplot}
\alias{ramplot}
\title{Convenience function to plot Raman spectra (from MALDIquant::MassSpectrum)}
\usage{
ramplot(
  msspc,
  main = "Raman spectrum",
  xlab = expression("Wavenumber (cm"^-1 * ")"),
  ylab = "Intensity (AU)",
  ...
)
}
\arguments{
\item{msspc}{a MALDIquant::MassSpectrum (trimmed to relevant wavelengths)}

\item{main}{a string with the plot header (defaults to "Raman Spectrum")}

\item{xlab}{a string with the horizontal axis label (defaults to inverse wavenumber)}

\item{ylab}{a string with the vertical axis label (devaults to Intensity (AU))}

\item{...}{additional parameters passed on to plot}
}
\description{
This convenience function is a wrapper around MALDIquant plot-methods.
In essence it does nothing else than rename the axes.
}
