% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_quiet.R
\name{run_quiet}
\alias{run_quiet}
\title{run_quiet function}
\usage{
run_quiet(x)
}
\arguments{
\item{x}{function call that needs to run quiet.}
}
\description{
Function to make other functions with cat() run quiet.
Credits to Hadley Wickham.
}
\keyword{run_quiet}
