% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.diss.R
\name{SCA.diss}
\alias{SCA.diss}
\title{Calculate the spectral contrast angle (SCA) between all cells in a hyperSpec object}
\usage{
SCA.diss(x)
}
\arguments{
\item{x}{a hyperSpec object containing the cells in the rows, and wavenumbers (features) in columns}
}
\value{
a distance object containing the SCA between each cell
}
\description{
The SCA measures the angle between two vectors corresponding to closely
 related spectra to measure whether they are the same or not
 (Wan *et al.* 2002). This function returns a dist object based on the SCA
 for downstream processing.
}
\examples{
## Short example

# Load and pretransform MassSpectrum object
data("mass.spectra.baseline.corr")
mass.spectra.baseline.corr <- wlcutter(mass.spectra.baseline.corr)
mq.norm <- calibrateIntensity(mass.spectra.baseline.corr, method="TIC",range=c(600, 1800))
# Convert to hyperSpec object
hs.norm <- mq2hs(mq.norm)
disst <- SCA.diss(hs.norm)
}
\references{
Wan, K. X., Vidavsky, I., & Gross, M. L. (2002). Comparing similar spectra: from similarity index to spectral contrast angle. Journal of the American Society for Mass Spectrometry, 13(1), 85-88.
}
