% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R
\name{SCA}
\alias{SCA}
\title{Calculate the spectral contrast angle (SCA) between two vectors}
\usage{
SCA(a, b)
}
\arguments{
\item{a}{vector 1}

\item{b}{vector 2}
}
\description{
The SCA measures the angle between two vectors corresponding to closely
 related spectra to measure whether they are the same or not
 (Wan *et al.* 2002).
}
\examples{
# Load hyperSpec object
data("hs_example")

# Convert to MassSpectrum object
hs.x.proc <- hs_preprocess(hs_example)

# Calc SCA
disst <- SCA(hs.x.proc@data$spc[1,], hs.x.proc@data$spc[2,])
}
\references{
Wan, K. X., Vidavsky, I., & Gross, M. L. (2002). Comparing similar spectra: from similarity index to spectral contrast angle. Journal of the American Society for Mass Spectrometry, 13(1), 85-88.
}
