% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spx.all.R
\docType{data}
\name{spx.all}
\alias{spx.all}
\title{Raman spectrometry test single-cell data of Escherichia coli LMG2092 biological replicates}
\format{
Hyperspec dataset with 4 slots
}
\usage{
spx.all
}
\description{
E. coli 2092 cultured in LB and NB, in biological triplicates at 28°C and
120 rpm on an orbital shaker.
After 24h, cells are measured using FCM and fixed with PFA.
They are measured that day (day1, all samples).
Check biological triplicates grown in LB and NB
}
\keyword{datasets}
