% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlcutter.R
\name{wlcutter}
\alias{wlcutter}
\title{Set intensities of a given bandpass to a preset value (anomalous peak removal)}
\usage{
wlcutter(x, lower_wl = 900L, upper_wl = 1100L, new_intens = 0L)
}
\arguments{
\item{x}{a list of MALDIquant::massSpectrum objects}

\item{lower_wl}{an integer of the lower boundary of wavelenghts to replace
intensities for (defaults to 900)}

\item{upper_wl}{an integer of the upper boundary of wavelengths to replace
intensities for (defaults to 1100)}

\item{new_intens}{new intensity value to assign to all peaks between lower_wl
and upper_wl. Defaults to 0. Needs to be a single integer value (for now).}
}
\value{
a list of  MALDIquant::massSpectrum objects with replaced intensities
}
\description{
This function converts takes a list of MALDIquant::massSpectrum objects and
sets the intensitiy values between certain wavelengths to a given value.
By default, this value is zero, as this function can be used to remove errant
peaks from the spectra.
}
\examples{
## Short example

# Load hyperSpec object
data("mass.spectra.baseline.corr")

# Convert to MassSpectrum object
massspec_newintens <- wlcutter(mass.spectra.baseline.corr)
}
