% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_contrast.R
\name{hs_contrast}
\alias{hs_contrast}
\title{Contrast plot for Raman spectra}
\usage{
hs_contrast(hs.x, comp1, comp2, plot = TRUE)
}
\arguments{
\item{hs.x}{a hyperSpec object}

\item{comp1}{numeric (any length) or boolean (TRUE/FALSE - length equal to number of spectra in hs.x) vector
with sample names from one side of the comparisons}

\item{comp2}{numeric (any length) or boolean (TRUE/FALSE - length equal to number of spectra in hs.x) vector
with sample names from the other side of the comparisons}

\item{plot}{a logical indicating whether or not the constrasts should be plotted (defaults to TRUE)}
}
\value{
a dataframe containing the contrast values between the averaged spectra
}
\description{
A way to visualize contrasts between Raman spectra from different samples in an
 existing hyperSpec object
}
\examples{
## Short example
data("hs_example")

# Preprocess
hs_example <- hs_preprocess(hs_example)

# Run contrast comparison
hs.cont <- hs_contrast(hs_example, comp1 = c(1:4), comp2 = c(8:20))
}
