% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_hclust.R
\name{hs_hclust}
\alias{hs_hclust}
\title{Perform hierarchical clustering on Raman spectra}
\usage{
hs_hclust(
  hs.x,
  dist_method = "bray",
  clust_method = "pvclust",
  aggl_method = "ward.D2",
  nboot = 1000,
  ...
)
}
\arguments{
\item{hs.x}{hyperSpec object}

\item{dist_method}{Dissimilarity index, partial match to "manhattan", "euclidean", "canberra", "clark",
"bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn",
"mountford", "raup", "binomial", "chao", "cao" or "mahalanobis". Can be specified to "SCA" for Raman data.}

\item{clust_method}{Choose regular hierarchical clustering or bootstrap supported hierarchical clustering with pvclust}

\item{aggl_method}{the agglomerative method used in hierarchical clustering. This should be (an abbreviation of) one of "average", "ward.D", "ward.D2", "single", "complete", "mcquitty", "median" or "centroid". The default is "average".}

\item{nboot}{Numnber of bootstraps for pvclust. Defaults to 1000.}

\item{...}{Parameters to pass on to hclust().}

\item{method.hclust}{Clustering method to use for pvclust. Either hclust or pvclust. Defaults to "pvclust"..}
}
\description{
Perform hierarchical clustering on Raman spectra
}
\examples{
## Short example

# Load hyperSpec object
data("hs_example")

# Preprocess spectra
hs.x.proc <- hs_preprocess(hs_example)
hs.x.proc <- hs_resample(hs.x.proc, sample = 10)

# Cluster
hclust_obj <- hs_hclust(hs.x.proc, dist_method = "bray",
clust_method = "hclust")
}
