% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_import.R
\name{hs_import}
\alias{hs_import}
\title{Import Raman spectral data in the R environment}
\usage{
hs_import(path = NULL)
}
\arguments{
\item{path}{path}
}
\description{
This function imports Raman data (Thermo Galactic's spc file format) into the R environment
}
\examples{
# Import Raman spectral data
hs.x <- hs_import(path = "")
}
