% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_phenoRam.R
\name{hs_phenoRam}
\alias{hs_phenoRam}
\title{Calculate the phenotypic heterogeneity indices for each cell}
\usage{
hs_phenoRam(
  hs.x,
  preprocess = FALSE,
  smooth = FALSE,
  align = FALSE,
  path = NULL,
  trim.range = c(400, 1800),
  pattern = ".spc",
  niter = 10,
  return_spectra = TRUE,
  peak_detection = FALSE,
  peak_window = 20,
  peak_method = c("MAD")
)
}
\arguments{
\item{hs.x}{d}

\item{preprocess}{d}

\item{smooth}{d}

\item{align}{d}

\item{niter}{d}

\item{return_spectra}{d}

\item{peak_detection}{d}

\item{peak_window}{d}

\item{peak_method}{d}

\item{d}{d}
}
\description{
Calculate the phenotypic heterogeneity indices for each cell
}
\examples{
# Short example
data("hs_example")

# Preprocess
hs_example <- hs_preprocess(hs_example)

# Calculate metrics
hs_phenoRam(hs.x = hs_example, preprocess = TRUE, peak_detection = TRUE)
}
