% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_preprocess.R
\name{hs_preprocess}
\alias{hs_preprocess}
\title{Calculate the phenotypic heterogeneity indices for each cell}
\usage{
hs_preprocess(
  hs.x,
  smooth = FALSE,
  align = FALSE,
  path = NULL,
  pattern = ".spc",
  trim.range = c(400, 1800),
  niter = 10
)
}
\arguments{
\item{hs.x}{Available hyperspec object}

\item{smooth}{Smoothing TRUE/FALSE}

\item{align}{Align spectra TRUE/FALSE}

\item{path}{Path contain spc files if no hyperSpec object is available}

\item{pattern}{Pattern of spc files to import from path}

\item{trim.range}{Wavenumber range to trim from raw spectra.}

\item{niter}{Number of iterations.}
}
\description{
Calculate the phenotypic heterogeneity indices for each cell
}
\examples{
# Load hyperSpec object
data("hs.example")

# Load hyperSpec object
hs.x.proc <- hs_preprocess(hs.x = hs.example, smooth = TRUE)
}
