% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterationsplot.R
\name{iterationsplot}
\alias{iterationsplot}
\title{Plot the effect of iterations on baseline correction}
\usage{
iterationsplot(mdq, itervect, plotlegend = TRUE, ...)
}
\arguments{
\item{mdq}{a MALDIquant::MassSpectrum (trimmed to relevant wavelengths)}

\item{itervect}{a numeric vector of different iteration options (max. 11 values)}

\item{plotlegend}{a boolean indicating wether or not to include a legend in lower left corner
(defaults to TRUE)}

\item{...}{additional parameters passed on to plot or lines

## Short example

# Load hyperSpec object
data("mdqs")

# Extract wavelengths and plot
mdqs.trim <- trim(mdqs, range=c(600, 1800))
selms <- mdqs.trim[[1]]
iteration.options <- c(5,10,20,30,40,50,100)
iterationsplot(mdq=selms,itervect=iteration.options)
# if you want to check for all
# pdf("iterationplots.pdf")
# lapply(mdqs,iterationsplot,iteration.options)
# dev.off()}
}
\description{
This function plots baseline correction for several iterations on the
spectral data to determine the optimal number of iterations
}
