% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA.R
\name{SCA}
\alias{SCA}
\title{Calculate the spectral contrast angle (SCA) between two vectors}
\usage{
SCA(a, b)
}
\arguments{
\item{a}{vector 1}

\item{b}{vector 2}
}
\description{
The SCA measures the angle between two vectors corresponding to closely
 related spectra to measure whether they are the same or not
 (Wan *et al.* 2002).
}
\examples{
## Short example

# Load and pretransform MassSpectrum object
data("mass.spectra.baseline.corr")
mass.spectra.baseline.corr <- wlcutter(mass.spectra.baseline.corr)
mq.norm <- calibrateIntensity(mass.spectra.baseline.corr, method="TIC",range=c(600, 1800))
# Convert to hyperSpec object
hs.norm <- mq2hs(mq.norm)
disst <- SCA(hs.norm@data$spc[1,],hs.norm@data$spc[2,])
}
\references{
Wan, K. X., Vidavsky, I., & Gross, M. L. (2002). Comparing similar spectra: from similarity index to spectral contrast angle. Journal of the American Society for Mass Spectrometry, 13(1), 85-88.
}
